+++ core/domain/src/main/java/com/mmrtr/lol/domain/league/application/usecase/CalculateSummonerRankingUseCase.java
@@ -1,4 +1,4 @@
-package com.mmrtr.lol.domain.league.service.usecase;
+package com.mmrtr.lol.domain.league.application.usecase;
 
 import com.mmrtr.lol.domain.league.domain.SummonerRanking;
 import com.mmrtr.lol.domain.league.repository.SummonerRankingRepositoryPort;
