+++ core/domain/src/main/java/com/mmrtr/lol/domain/league/application/usecase/TriggerSummonerRankingCalculationUseCase.java
@@ -1,4 +1,4 @@
-package com.mmrtr.lol.domain.league.service.usecase;
+package com.mmrtr.lol.domain.league.application.usecase;
 
 public interface TriggerSummonerRankingCalculationUseCase {
     void execute();
