+++ infra/rabbitmq/src/main/java/com/mmrtr/lol/infra/rabbitmq/listener/MatchListener.java
@@ -2,8 +2,8 @@ package com.mmrtr.lol.infra.rabbitmq.listener;
 
 import com.mmrtr.lol.domain.match.readmodel.MatchDto;
 import com.mmrtr.lol.domain.match.readmodel.timeline.TimelineDto;
-import com.mmrtr.lol.domain.match.service.port.MatchApiPort;
-import com.mmrtr.lol.domain.match.service.usecase.SaveMatchDataUseCase;
+import com.mmrtr.lol.domain.match.application.port.MatchApiPort;
+import com.mmrtr.lol.domain.match.application.usecase.SaveMatchDataUseCase;
 import com.mmrtr.lol.infra.rabbitmq.config.RabbitMqBinding;
 import com.rabbitmq.client.Channel;
 import lombok.RequiredArgsConstructor;
