+++ infra/rabbitmq/src/main/java/com/mmrtr/lol/infra/rabbitmq/service/SummonerRenewalService.java
@@ -3,9 +3,9 @@ package com.mmrtr.lol.infra.rabbitmq.service;
 import com.mmrtr.lol.common.type.Platform;
 import com.mmrtr.lol.domain.match.readmodel.MatchDto;
 import com.mmrtr.lol.domain.match.readmodel.timeline.TimelineDto;
-import com.mmrtr.lol.domain.match.service.usecase.SaveMatchDataUseCase;
+import com.mmrtr.lol.domain.match.application.usecase.SaveMatchDataUseCase;
 import com.mmrtr.lol.domain.summoner.domain.Summoner;
-import com.mmrtr.lol.domain.summoner.service.usecase.SaveSummonerDataUseCase;
+import com.mmrtr.lol.domain.summoner.application.usecase.SaveSummonerDataUseCase;
 import com.mmrtr.lol.infra.rabbitmq.config.RabbitMqBinding;
 import com.mmrtr.lol.infra.rabbitmq.dto.SummonerRenewalMessage;
 import com.mmrtr.lol.infra.rabbitmq.service.MatchDataFetcher.FetchNewMatchIdsResult;
