+++ infra/riot-client/src/main/java/com/mmrtr/lol/infra/riot/adapter/MatchApiAdapter.java
@@ -3,7 +3,7 @@ package com.mmrtr.lol.infra.riot.adapter;
 import com.mmrtr.lol.common.type.Platform;
 import com.mmrtr.lol.domain.match.readmodel.MatchDto;
 import com.mmrtr.lol.domain.match.readmodel.timeline.TimelineDto;
-import com.mmrtr.lol.domain.match.service.port.MatchApiPort;
+import com.mmrtr.lol.domain.match.application.port.MatchApiPort;
 import com.mmrtr.lol.infra.riot.service.RiotApiService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
